CB-8908 Make fetching via git faster via --depth=1

+++ cordova-lib/src/gitclone.js
@@ -21,8 +21,8 @@ var  Q             = require('q'),
      shell         = require('shelljs'),
      events        = require('./events'),
      path          = require('path'),
-     globalUtil    = require('util'),
-     child_process = require('child_process'),
+     semver        = require('semver'),
+     superspwan    = require('./cordova/superspawn'),
      os            = require('os');
 
 
@@ -31,6 +31,7 @@ exports.clone = clone;
 //  clone_dir, if provided is the directory that git will clone into.
 //  if no clone_dir is supplied, a temp directory will be created and used by git.
 function clone(git_url, git_ref, clone_dir){
+    var needsGitCheckout = !!git_ref;
     if (!shell.which('git')) {
         return Q.reject(new Error('"git" command line tool is not installed: make sure it is accessible on your PATH.'));
     }
@@ -43,19 +44,31 @@ function clone(git_url, git_ref, clone_dir){
     shell.rm('-rf', tmp_dir);
     shell.mkdir('-p', tmp_dir);
 
-    var cmd = globalUtil.format('git clone "%s" "%s"', git_url, tmp_dir);
-    events.emit('verbose', 'Cloning repository via git-clone command: ' + cmd);
-
-    return Q.ninvoke(child_process, 'exec', cmd, {}).then(function () {
-	if(git_ref){
-	    var checkoutCmd = globalUtil.format('git checkout "%s"', git_ref);
-            events.emit('verbose', 'Checking out git ref via command: ' + checkoutCmd);
-            return Q.ninvoke(child_process, 'exec', checkoutCmd, { cwd: tmp_dir }).then(function(){
-		events.emit('log', 'Repository "' + git_url + '" checked out to git ref "' + git_ref + '".');
-	    });
+    return superspwan.spawn('git', ['--version'])
+    .then(function(output) {
+        var gitVersion = /\d+\.\d+(\.\d+)?/.exec(output);
+        gitVersion = gitVersion ? gitVersion[0] : '1.0.0';
+        var cloneArgs = ['clone'];
+        if (semver.gte(gitVersion, '1.7.0')) {
+            if (git_ref) {
+                cloneArgs.push('--branch', git_ref);
+                needsGitCheckout = false;
+            }
+            if (semver.gte(gitVersion, '1.7.10')) {
+                cloneArgs.push('--single-branch');
+            }
+            cloneArgs.push('--depth=1');
+        }
+        cloneArgs.push('--', git_url, tmp_dir);
+        return superspwan.spawn('git', cloneArgs);
+    }).then(function() {
+	if (needsGitCheckout){
+            return superspwan.spawn('git', ['checkout', git_ref], {
+                cwd: tmp_dir
+            });
 	}
-	return Q();
     }).then(function(){
+        events.emit('log', 'Repository "' + git_url + '" checked out to git ref "' + git_ref + '".');
 	return tmp_dir;
     }).fail(function (err) {
         shell.rm('-rf', tmp_dir);
